code shuffling to avoid circular import
